requirements.txt
streamlit
pdfminer.six
google-generativeai